+++ src/Pulse/Program.cs
@@ -37,17 +37,17 @@ internal class Program {
             return await cli.RunAsync(args, false);
         } catch (Exception e) when (e is TaskCanceledException or OperationCanceledException) {
             GoToLine(firstLine);
-            ClearNextLines(4);
-            ClearNextLinesError(4);
+            ClearNextLines(4, OutputPipe.Out);
+            ClearNextLines(4, OutputPipe.Error);
             WriteLine("Cancellation requested and handled gracefully." * Color.DarkYellow);
             return 1;
         } catch (Exception e) {
             GoToLine(firstLine);
-            ClearNextLines(4);
-            ClearNextLinesError(4);
-            WriteLineError("Unexpected error! Contact developer at dusrdev@gmail.com" * Color.Red);
-            WriteLineError("And provide the following output:" * Color.Red);
-            NewLine();
+            ClearNextLines(4, OutputPipe.Out);
+            ClearNextLines(4, OutputPipe.Error);
+            WriteLine("Unexpected error! Contact developer at dusrdev@gmail.com" * Color.Red, OutputPipe.Error);
+            WriteLine("And provide the following output:" * Color.Red, OutputPipe.Error);
+            NewLine(OutputPipe.Error);
             Helper.PrintException(StrippedException.FromException(e));
             return 1;
         }