+++ src/Pulse/Program.cs
@@ -7,7 +7,7 @@ using static PrettyConsole.Console;
 using PrettyConsole;
 
 internal class Program {
-	internal const string VERSION = "1.0.5.0";
+    internal const string VERSION = "1.0.6.0";
 
     private static async Task<int> Main(string[] args) {
         using CancellationTokenSource globalCTS = new();
@@ -37,17 +37,18 @@ internal class Program {
             return await cli.RunAsync(args, false);
         } catch (Exception e) when (e is TaskCanceledException or OperationCanceledException) {
             GoToLine(firstLine);
-            ClearNextLines(4);
-            ClearNextLinesError(4);
+            ClearNextLines(4, OutputPipe.Out);
+            ClearNextLines(4, OutputPipe.Error);
             WriteLine("Cancellation requested and handled gracefully." * Color.DarkYellow);
             return 1;
         } catch (Exception e) {
             GoToLine(firstLine);
-            ClearNextLines(4);
-            ClearNextLinesError(4);
-            WriteLineError("Unexpected error! Contact developer and provide the following output:" * Color.Red);
-            NewLine();
-            WriteLine(JsonContext.SerializeException(e));
+            ClearNextLines(4, OutputPipe.Out);
+            ClearNextLines(4, OutputPipe.Error);
+            WriteLine("Unexpected error! Contact developer at dusrdev@gmail.com" * Color.Red, OutputPipe.Error);
+            WriteLine("And provide the following output:" * Color.Red, OutputPipe.Error);
+            NewLine(OutputPipe.Error);
+            Helper.PrintException(StrippedException.FromException(e));
             return 1;
         }
     }