+++ Changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Version 1.0.6.0
+
+- Updated to .NET 9.0
+- Optimized using new version of `PrettyConsole` (3.1.0)
+
 ## Version 1.0.5.0
 
 - Request duration is was changed to latency, a more important metric to measure